fix: extract shared ClaimPendedEvent contract to prevent schema drift between producer and consumer
