+++ cloudhealthoffice-main.sln
@@ -175,6 +175,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "claims-examiner-service", "
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudHealthOffice.ClaimsExaminerService.Tests", "tests\CloudHealthOffice.ClaimsExaminerService.Tests\CloudHealthOffice.ClaimsExaminerService.Tests.csproj", "{63F7AAFC-0279-4537-AD38-5D188D73739C}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Shared", "Shared", "{F078A44D-36FD-4053-8C1D-0DA235C6B161}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CloudHealthOffice.Events", "src\shared\CloudHealthOffice.Events\CloudHealthOffice.Events.csproj", "{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -617,6 +621,18 @@ Global
 		{1C6BBE9C-F722-4FBB-8A8E-30C51162A6D4}.Release|x64.Build.0 = Release|Any CPU
 		{1C6BBE9C-F722-4FBB-8A8E-30C51162A6D4}.Release|x86.ActiveCfg = Release|Any CPU
 		{1C6BBE9C-F722-4FBB-8A8E-30C51162A6D4}.Release|x86.Build.0 = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|x64.Build.0 = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Debug|x86.Build.0 = Debug|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|x64.ActiveCfg = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|x64.Build.0 = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|x86.ActiveCfg = Release|Any CPU
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68}.Release|x86.Build.0 = Release|Any CPU
 		{743FA11B-0D41-492E-8619-C63CF0DD64AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{743FA11B-0D41-492E-8619-C63CF0DD64AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{743FA11B-0D41-492E-8619-C63CF0DD64AE}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -1071,6 +1087,8 @@ Global
 		{3F92B1C3-DBA6-4EB3-88E4-B7B24296A241} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
 		{0C3DE287-12F2-4127-8A43-CD77E7FB5001} = {984BB9B3-3FA3-BE33-9484-CAC21695A33C}
 		{63F7AAFC-0279-4537-AD38-5D188D73739C} = {0AB3BF05-4346-4AA6-1389-037BE0695223}
+		{F078A44D-36FD-4053-8C1D-0DA235C6B161} = {827E0CD3-B72D-47B6-A68D-7590B98EB39B}
+		{8DC1EBCE-63CF-42DE-8313-D82F4B949B68} = {F078A44D-36FD-4053-8C1D-0DA235C6B161}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {C10C3349-5BEE-4B34-9987-23F0142F9D98}
