+++ src/services/claims-examiner-service/Models/ClaimSnapshot.cs
@@ -1,3 +1,5 @@
+using CloudHealthOffice.Events;
+
 namespace ClaimsExaminerService.Models;
 
 /// <summary>
