+++ src/services/claims-examiner-service/Services/Examiner/ExaminerOrchestrator.cs
@@ -1,6 +1,7 @@
 using System.Text.Json.Nodes;
 using ClaimsExaminerService.Models;
 using ClaimsExaminerService.Services.Anthropic;
+using CloudHealthOffice.Events;
 
 namespace ClaimsExaminerService.Services.Examiner;
