+++ src/services/claims-examiner-service/Services/Examiner/ExaminerPromptBuilder.cs
@@ -2,6 +2,7 @@ using System.Text;
 using System.Text.Json.Nodes;
 using ClaimsExaminerService.Models;
 using ClaimsExaminerService.Services.Anthropic;
+using CloudHealthOffice.Events;
 
 namespace ClaimsExaminerService.Services.Examiner;
