+++ src/services/claims-service/claims-service.csproj
@@ -9,6 +9,7 @@
 
   <ItemGroup>
     <ProjectReference Include="../shared/CloudHealthOffice.Infrastructure/CloudHealthOffice.Infrastructure.csproj" />
+    <ProjectReference Include="../../shared/CloudHealthOffice.Events/CloudHealthOffice.Events.csproj" />
   </ItemGroup>
 
   <ItemGroup>
