+++ src/shared/CloudHealthOffice.Events/CloudHealthOffice.Events.csproj
@@ -1,0 +1,11 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <RootNamespace>CloudHealthOffice.Events</RootNamespace>
+    <AssemblyName>CloudHealthOffice.Events</AssemblyName>
+  </PropertyGroup>
+
+</Project>
